+++ src/HttpRequestDurationsMiddlewareBuilderExtensions.cs
@@ -29,12 +29,6 @@ namespace Prometheus.Client.HttpRequestDurations
                 ??= (ICollectorRegistry)app.ApplicationServices.GetService(typeof(ICollectorRegistry))
                     ?? Metrics.DefaultCollectorRegistry;
 
-#if HasRoutes
-            // If we are going to set labels for Controller or Action -- then we need to make them readily available
-            if (options.IncludeController || options.IncludeAction)
-                app.UseMiddleware<CaptureRouteDataMiddleware>();
-#endif
-
             return app.UseMiddleware<HttpRequestDurationsMiddleware>(options);
         }
     }
